AI4DM_H2_Milestone 1_Viswak Raja/requirements.txt
flask
python-dotenv
ipykernel
torch
pillow
transformers
openai
diffusers
matplotlib
accelerate